+++ src/main/kotlin/com/bartek/clinicadministrationapi/services/VisitService.kt
@@ -3,9 +3,11 @@ package com.bartek.clinicadministrationapi.services
 import com.bartek.clinicadministrationapi.domain.dtos.VisitDTO
 import com.bartek.clinicadministrationapi.mappers.VisitMapper
 import com.bartek.clinicadministrationapi.repositories.VisitRepository
+import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
+import java.util.*
 
 @Service
 class VisitService(
@@ -15,57 +17,66 @@ class VisitService(
 
     fun addVisit(visitDTO: VisitDTO): ResponseEntity<VisitDTO> {
 
-        val isTrue: VisitDTO? = visitDTO.doctor.id?.let {
+        val visitDb: VisitDTO? = visitDTO.doctor.id?.let {
             visitRepository.findVisitsByDateAndDateTimeAndDoctorId(visitDTO.date, visitDTO.dateTime, it)
         }?.let { visitMapper.mapDAOToDTO(it) }
 
-        return if (isTrue != null) {
-            ResponseEntity(HttpStatus.BAD_REQUEST)
+        if (visitDb != null) {
+            return ResponseEntity(HttpStatus.BAD_REQUEST)
         } else {
             visitRepository.save(visitMapper.mapDTOToDAO(visitDTO))
-            ResponseEntity(HttpStatus.OK)
+            return ResponseEntity(HttpStatus.OK)
         }
     }
 
-    fun getAllVisitsByPatientId(id: Long): ResponseEntity<Set<VisitDTO>> {
-        val set: Set<VisitDTO> = visitRepository.findVisitsByPatientId(id)
+    fun getAllVisitsByPatientId(id: Long): Optional<Set<VisitDTO>> {
+
+        val optVisit = Optional.of(visitRepository.findVisitsByPatientId(id)
             .map { visitDAO -> visitMapper.mapDAOToDTO(visitDAO) }
-            .toSet()
+            .toSet())
 
-        return ResponseEntity(set, HttpStatus.OK)
-
-    }
-
-    fun deleteVisitById(id: Long): ResponseEntity<VisitDTO> {
-        visitRepository.deleteById(id)
-        return ResponseEntity(HttpStatus.OK)
-    }
-
-    fun updateVisit(visitDTO: VisitDTO): ResponseEntity<VisitDTO>? {
-
-        val isTrue: VisitDTO? =
-            visitDTO.doctor.id?.let {
-                visitRepository.findVisitsByDateAndDateTimeAndDoctorId(visitDTO.date, visitDTO.dateTime, it)
-                    ?.let { visitMapper.mapDAOToDTO(it) }
-            }
-
-        if (isTrue != null) {
-            return ResponseEntity(HttpStatus.BAD_REQUEST)
+        if (optVisit.isPresent) {
+            return optVisit
         } else {
+            throw EmptyResultDataAccessException("Visit not found by patient id", 0)
+        }
+    }
+
+    fun deleteVisitById(id: Long): Optional<VisitDTO> {
+
+        val visitOpt = visitRepository.findById(id)
+            .map { visitDAO -> visitMapper.mapDAOToDTO(visitDAO) }
+
+        if (visitOpt.isPresent) {
+            visitRepository.deleteById(id)
+            return visitOpt
+        } else {
+            throw EmptyResultDataAccessException("Visit not found by this id", 0)
+        }
+    }
+
+    fun updateVisit(visitDTO: VisitDTO): Optional<VisitDTO>? {
+
+        val visitDb = visitDTO.doctor.id?.let {
+            visitRepository.findVisitsByDateAndDateTimeAndDoctorId(visitDTO.date, visitDTO.dateTime, it)
+        }?.let { visitMapper.mapDAOToDTO(it) }?.let { Optional.of(it) }
+
+        if (visitDb != null) {
             visitDTO.id?.let {
                 visitRepository.findById(it)
                     .map { visitRepository.save(visitMapper.mapDTOToDAO(visitDTO)) }
             }
-            return ResponseEntity(HttpStatus.OK)
         }
+        return visitDb
 
     }
 
-    fun getAllVisits(): ResponseEntity<Set<VisitDTO>> {
-        val set: Set<VisitDTO> = visitRepository.findAll()
-            .map { visitDAO -> visitMapper.mapDAOToDTO(visitDAO) }
-            .toSet()
+    fun getAllVisits(): Optional<Set<VisitDTO>> {
 
-        return ResponseEntity(set, HttpStatus.OK)
+        val optSet = Optional.of(visitRepository.findAll()
+            .map { visitDAO -> visitMapper.mapDAOToDTO(visitDAO) }
+            .toSet())
+
+        return optSet
     }
 }
